+++ src/app/login/page.tsx
@@ -1,12 +1,12 @@
-import GithubLoginButton from "@/components/login/githubLoginButton/GithubLoginButton";
-import LoginDescription from "@/components/login/loginDescription/LoginDescription";
+import GithubLoginButton from "@/components/login/githubLoginButton/GithubLoginButton/GithubLoginButton";
+import LoginPageIntro from "@/components/login/logindescription/LoginPageIntro";
 import "./page.scss";
 
 const login =() =>{
     return(
         <div className="login-page">
             <div className="login-space-1"></div>
-            <LoginDescription/>
+            <LoginPageIntro/>
             <div className="login-space-2"></div>
             <GithubLoginButton/>
         </div>
