+++ src/components/description/TILRecordDescription/TILRecordDescription.tsx
@@ -6,9 +6,9 @@ import useUserInfoStore from '@/store/userInfoStore';
 const TILRecordDescription = () => {
   const { userInfo } = useUserInfoStore();
   return (
-    <div className="description">
+    <article className="description">
       {userInfo.name}님의 TIL 기록입니다.
-    </div>
+    </article>
   );
 };
