--- src/components/login/githubLoginButton/GithubLoginButton.scss
@@ -1,3 +1,0 @@
-.github-login-button {
-    cursor: pointer;
-  }
