+++ src/components/login/githubLoginButton/GithubLoginButton/GithubLoginButton.tsx
@@ -2,13 +2,9 @@
 
 import { useEffect } from 'react';
 import { useSearchParams } from 'next/navigation';
-import Image from 'next/image';
-import gitlogin from 'public/gitlogin.png';
-import './GithubLoginButton.scss';
-
 import GithubLogin from '@/api/login/loginAPI';
 import useAuthStore from '@/store/authStore';
-import { getGithubAuthUrl } from '@/api/constant/githubLoginConstants';
+import LoginButton from '../LoginButton/LoginButton';
 
 const GithubLoginButton = () => {
   const searchParams = useSearchParams();
@@ -19,8 +15,7 @@ const GithubLoginButton = () => {
     if (code) {
       GithubLogin(code)
         .then((data) => {
-          const token = data.accessToken;
-          setAccessToken(token);
+          setAccessToken(data.accessToken);
           window.location.href = '/';
         })
         .catch((error) => {
@@ -29,23 +24,7 @@ const GithubLoginButton = () => {
     }
   }, [searchParams, setAccessToken]);
 
-  const handleLogin = () => {
-    const client_id = process.env.NEXT_PUBLIC_GITHUB_CLIENT_ID!;
-    const githubAuthUrl = getGithubAuthUrl(client_id);
-    window.location.href = githubAuthUrl;
-  };
-
-  return (
-    <Image
-      onClick={handleLogin}
-      src={gitlogin}
-      alt="GitHub 로그인 버튼"
-      width={325}
-      height={45}
-      priority
-      className="github-login-button"
-    />
-  );
+  return <LoginButton />;
 };
 
 export default GithubLoginButton;
