+++ src/components/login/githubLoginButton/LoginButton/LoginButton.scss
@@ -1,0 +1,8 @@
+button {
+    background: none;
+    border: none;
+    padding: 0;
+    margin: 0;
+    cursor: pointer;
+  }
+
