+++ src/components/login/logindescription/LoginPageIntro.tsx
@@ -1,17 +1,17 @@
 import logindescription from "public/logindescription.png"
 import Image from "next/image";
-const LoginDescription =() =>{
+const LoginPageIntro =() =>{
     return(
         <div>
             <Image
             src={logindescription}
             alt="YouTIL Login"
             width={600}
-            height={180}
+            height={165}
             priority
             />
       </div>
     )
 }
 
-export default LoginDescription;
+export default LoginPageIntro;
